% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSubtypes.R
\name{addSubtypes}
\alias{addSubtypes}
\title{Add cancer subtypes to a cohort}
\usage{
addSubtypes(cohort, cdm, cancer, window = list(c(0, 0)), showIntersect = FALSE)
}
\arguments{
\item{cohort}{A cohort table with cancer patients from a cdm reference object.}

\item{cdm}{A cdm reference object.}

\item{cancer}{In character, the affected site, a choice of:
"bladder", "breast", "colorectal", "lung", "melanoma", "oesophagus" and "prostate".}

\item{window}{to look up stages codes.}

\item{showIntersect}{If TRUE, the cohort will show the date intersects for each matching code. Default FALSE.}
}
\description{
This function filters and appends specific cancer subtype information to the
cohort data based on the provided cancer site, timeframe window, and intersection logic.
}
