% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertCancerCohortName}
\alias{assertCancerCohortName}
\title{Assert that cohort names match selected cancer site(s)}
\usage{
assertCancerCohortName(cohort, cancer = "breast")
}
\arguments{
\item{cohort}{A cohort table whose names are to be validated.}

\item{cancer}{A character vector of supported cancer-site names, such as
\code{"breast"}.}
}
\value{
\code{NULL}, invisibly, if validation succeeds.
}
\description{
Validates that all names in a cohort table correspond to the selected cancer
site(s). Cohort names must follow the \code{"<cancer>_cancer"} format.
}
\details{
Throws an error of class \code{"Invalid cohort names"} if any cohort name does not
correspond to the selected cancer site(s).
}
