% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assertCharacteristic}
\alias{assertCharacteristic}
\title{Assert that a character vector contains only allowed characteristics}
\usage{
assertCharacteristic(x, characteristics)
}
\arguments{
\item{x}{A character vector to validate.}

\item{characteristics}{A character vector of allowed characteristic names.}
}
\value{
\code{NULL}, invisibly, if validation succeeds.
}
\description{
Validates that all elements in the input vector are present in the
provided list of allowed characteristics.
}
\details{
Throws an error of class \code{"Invalid characteristics"} if any element in \code{x}
is not present in \code{characteristics}.
}
\keyword{internal}
