% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{createTNMCodelist}
\alias{createTNMCodelist}
\title{Create a TNM codelist from concept data}
\usage{
createTNMCodelist(tnm_concepts, .edition, .type)
}
\arguments{
\item{tnm_concepts}{A data frame of packaged TNM concept definitions.}

\item{.edition}{A single staging-system edition used to filter concepts.}

\item{.type}{A single stage-rule grouping used to filter concepts.}
}
\value{
An \code{omopgenerics} codelist of TNM concepts.
}
\description{
Create a TNM codelist from concept data
}
\keyword{internal}
