% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{.addStageRules}
\alias{.addStageRules}
\title{Add stage rules to a cohort}
\usage{
.addStageRules(
  cohort,
  conceptSet,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = list(c(0, 0)),
  targetDate = "event_start_date",
  order = "last",
  inObservation = TRUE,
  nameStyle = "{concept_name}",
  name = NULL,
  ruleset
)
}
\arguments{
\item{cohort}{A cohort table.}

\item{conceptSet}{An \code{omopgenerics} codelist object.}

\item{indexDate}{A column in \code{cohort} used as the lookup-date origin.}

\item{censorDate}{An optional column in \code{cohort} that censors lookup events.}

\item{window}{A list of two-element numeric lookup windows in days.}

\item{targetDate}{The event-date column used for the intersection.}

\item{order}{Whether the first or last matching event is selected.}

\item{inObservation}{Whether to limit lookup to the observation period.}

\item{nameStyle}{A glue specification used for generated column names.}

\item{name}{An optional name for the generated table.}

\item{ruleset}{A data frame that maps T, N, and M combinations to stages.}
}
\value{
A cohort table with applied stage rules.
}
\description{
Add stage rules to a cohort
}
\keyword{internal}
