% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{.mapStageRules}
\alias{.mapStageRules}
\title{Map stage rules to a cohort}
\usage{
.mapStageRules(cohort, ruleset)
}
\arguments{
\item{cohort}{A cohort table containing intersected TNM values.}

\item{ruleset}{A data frame containing the stage rules.}
}
\value{
A cohort table with a new \code{cancer_stage} column.
}
\description{
Map stage rules to a cohort
}
\keyword{internal}
