% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{filterStageConcepts}
\alias{filterStageConcepts}
\title{Filter stage concepts from a codelist}
\usage{
filterStageConcepts(codelist)
}
\arguments{
\item{codelist}{An \code{omopgenerics} codelist containing stage concepts.}
}
\value{
A codelist containing only parent TNM stage concepts.
}
\description{
Filter stage concepts from a codelist
}
\keyword{internal}
