% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDuplicates.R
\name{shinyConceptDuplicates}
\alias{shinyConceptDuplicates}
\title{Visualize interactive UpSet plot of staging codelists with duplicates}
\usage{
shinyConceptDuplicates(cdm, concept_folder = "stages")
}
\arguments{
\item{cdm}{A CDM reference used to extract concept sets.}

\item{concept_folder}{A single folder name under \code{inst/concept_sets} containing
concept-set JSON files. Defaults to \code{"stages"}.}
}
\value{
Launches a Shiny app.
}
\description{
An UpSet plot is a good alternative to a Venn diagram to visualize intersections
of more than \eqn{3} sets: it shows which sets have common elements with which other set
and the size of the intersection, as well as the original size of every set.

The function \code{shinyConceptDuplicates()} launches a simple shiny app with filters for subcategory,
edition and classification to display a dynamic UpSet plot of staging codelists intersection.
}
