% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptIntersection.R
\name{shinyConceptIntersection}
\alias{shinyConceptIntersection}
\title{Visualize interactive UpSet plot of staging codelists intersections}
\usage{
shinyConceptIntersection(cdm, concept_folder = "stages")
}
\arguments{
\item{cdm}{A cdm instance, needed to extract concept sets.}

\item{concept_folder}{The character string indicating the folder under \code{inst/concept_sets}
where the concept sets of interest are saved in json files, default \code{"stages"}.}
}
\value{
Launches a Shiny app.
}
\description{
An UpSet plot is a good alternative to a Venn diagram to visualize intersections
of more than \eqn{3} sets: it shows which sets have common elements with which other set
and the size of the intersection, as well as the original size of every set.

The function \code{shinyConceptIntersection()} launches a simple shiny app with filters for subcategory,
edition and classification to display a dynamic UpSet plot of staging codelists intersection.
}
