% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{supportedCancerSites}
\alias{supportedCancerSites}
\title{Get supported cancer sites}
\usage{
supportedCancerSites()
}
\value{
A character vector of supported cancer-site names.
}
\description{
Get supported cancer sites
}
\keyword{internal}
