% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{supportedEdition}
\alias{supportedEdition}
\title{Get supported editions}
\usage{
supportedEdition()
}
\value{
A character vector of supported staging-system editions.
}
\description{
Get supported editions
}
\keyword{internal}
