% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{supportedType}
\alias{supportedType}
\title{Get supported type groupings}
\usage{
supportedType()
}
\value{
A character vector of supported stage-rule groupings.
}
\description{
Get supported type groupings
}
\keyword{internal}
