% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagesRules.R
\name{tnm_rules_docs}
\alias{tnm_rules_docs}
\title{Documentation of TNM staging rules}
\source{
\itemize{
\item Sobin, L. H., Gospodarowicz, M. K., & Wittekind, C. (Eds.). (2009).
TNM classification of malignant tumours (7th ed.). Wiley-Blackwell.
\item Brierley, J. D., Gospodarowicz, M. K., & Wittekind, C. (Eds.). (2016).
TNM classification of malignant tumours (8th ed.). Wiley-Blackwell.
\item Brierley, J. D., Giuliani, M., O’Sullivan, B., Rous, B., &
Van Eycken, E. (Eds.). (2025). TNM classification of malignant tumours
(9th ed.). Wiley.
}
}
\description{
Internal auxiliary files used to determine cancer stage.
}
\details{
The package stores three rule sets derived from UICC guidelines for the
TNM staging system. The rules stored in these files are differentiated by
cancer \strong{site} (bladder, breast, colorectal, lung, oesophageal, prostate,
skin), classification \strong{type} (clinical, pathological, base) and
classification \strong{edition} (7th, 8th, unspecified). The files are:

\enumerate{
\item \code{tnm_concepts.rds}: contains information about the individual TNM
components and their concept ids, differentiated by type and edition.
There are \eqn{393} total concepts for \eqn{44} unique components:

\if{html}{\out{<div class="sourceCode">}}\preformatted{TX, T0, Tis, Ta, T1, T1a, T1b, T1c, T1mi, T2, T2a, T2b, T2c, T3, T3a,
T3b, T4, T4a, T4b, T4c, T4d, NX, N0, N1, N1a, N1b, N1c, N1mi, N2, N2a,
N2b, N2c, N3, N3a, N3b, N3c, M0, M1, M1a, M1b, M1c, M1c1, M1c2, M1d.
}\if{html}{\out{</div>}}

These components have different versions according to:
\itemize{
\item Edition: \eqn{131} concepts for 7th, \eqn{131} for 8th, \eqn{131}
for unspecified;
\item Type: \eqn{132} concepts for base, \eqn{132} for clinical, \eqn{129}
for pathological (the \code{M0} component is not valid in the pathological
setting, for any of the editions).
}

\item \code{tnm_mapping.rds}: contains the rules to determine the cancer
stage based on a combination of individual TNM components, differentiated
by cancer site, type (clinical, pathological, base) and by edition (7th,
8th, 9th). Each rule refers to a specific source page of the UICC guidelines.

The currently available rules support the following concept categories.

For \code{bladder} cancer:
\itemize{
\item Edition: 7th, 8th, 9th available;
\item Type: only "base" is available, for all editions.
}
For \code{breast} cancer:
\itemize{
\item Edition: 7th, 8th, 9th available;
\item Type: only "base" is available, for all editions.
}
For \code{colorectal} cancer:
\itemize{
\item Edition: 7th, 8th, 9th available;
\item Type: only "base" is available, for all editions.
}
For \code{lung} cancer:
\itemize{
\item Edition: 7th, 8th, 9th available;
\item Type: only "base" is available, for all editions.
}
For \code{oesophageal} cancer:
\itemize{
\item For 7th edition, only "base" type is available;
\item For 8th edition, only "clinical" and "pathological" types are available but not "base".
\item For 9th edition, only "clinical" and "pathological" types are available but not "base".
}
For \code{prostate} cancer:
\itemize{
\item For 7th edition, only "base" type is available;
\item For 8th edition, only "clinical" type is available;
\item For 9th edition, only "clinical" and "pathological" types are available but not "base".
}
For \code{skin} cancer:
\itemize{
\item For 7th edition, only "pathological" type is available;
\item For 8th edition, only "clinical" and "pathological" types are available but not "base".
\item For 9th edition, only "clinical" and "pathological" types are available but not "base".
}

}

These files are meant for internal use and are not intended
to be modified by users.
}
\keyword{internal}
