% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{addStages}
\alias{addStages}
\title{\code{addStages()} to a cohort}
\usage{
addStages(
  cohort,
  cdm,
  cancer,
  window = list(c(0, 0)),
  edition = "8th",
  type = "base",
  order = "last",
  showTnm = FALSE
)
}
\arguments{
\item{cohort}{A cohort table with cancer patients from a
cdm reference object.}

\item{cdm}{A cdm reference object.}

\item{cancer}{In character, the affected site, a choice of:
"bladder", "breast", "colorectal", "lung", "melanoma", "oesophagus"
and "prostate".}

\item{window}{to look up stages codes.}

\item{edition}{A choice of "unspecified", "7th" and "8th".}

\item{type}{A choice from "base", "clinical" or "pathological" stage rule.}

\item{order}{A choice from "first" or "last". If more than one code
intersected, the order defines which code to intersect in the window.}

\item{showTnm}{If TRUE, the cohort will show the date intersects
for each matching code. Default FALSE.}
}
\value{
A cohort table containing the identified cancer stages.
}
\description{
It uses a codelist to date intersect with a cancer cohort.
Imposes a predefined or custom set of rules to identify
summary stages.
}
