% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStages.R
\name{addStages}
\alias{addStages}
\title{addCancerStages() information to a cohort}
\usage{
addStages(
  cohort,
  cdm,
  cancer,
  window = c(0, 0),
  edition = "eight",
  type = "base",
  order = "last"
)
}
\arguments{
\item{cohort}{A cohort table from a cdm reference object.}

\item{cdm}{A cdm reference.}

\item{stageConcepts}{A concept-set list containing cancer
stage concepts.}

\item{ruleSet}{A set of rules in list format that
corresponds to each element of the stageConcepts codelist.}
}
\value{
A cohort table containing the identified cancer stages.
}
\description{
It uses a codelist to date intersect with a cancer cohort.
Imposes a predefined or custom set of rules to identify
summary stages.
}
