% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptIntersection.R
\name{checkConceptIntersection}
\alias{checkConceptIntersection}
\title{Check if concept sets have common elements}
\usage{
checkConceptIntersection(cdm, concept_folder = "stages")
}
\arguments{
\item{cdm}{A cdm instance, needed to extract concept sets.}

\item{concept_folder}{The character string indicating the folder under \code{inst/concept_sets}
where the concept sets of interest are saved in json files, default \code{stages}.}
}
\value{
\code{TRUE} if the codelists have any elements in common, \code{FALSE} if they don't.
}
\description{
This function performs a simple boolean check to determine whether the codelists
generated from a number of concept sets have any element in common.
}
\details{
The intersection of codelists is calculated as follows: given an array of codelists,
the function uses the outer product of the array with the array itself to calculate
the length of the intersection of every possible combination of two codelists.
The result is a symmetric matrix where the element in position \eqn{(i, j)} is the number
of common codes between codelist \eqn{i} and codelist \eqn{j}, and the elements on
the diagonal represent the size of every codelist.

To simply assess the presence of common codes, it is enough to set the diagonal to \eqn{0}
and check if there is any other element \eqn{> 0} in the matrix.
}
