% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.R
\name{createCancerCohorts}
\alias{createCancerCohorts}
\title{Create cancer cohorts based on phenotype specifications}
\usage{
createCancerCohorts(cdm, path = "cancer_cohorts", name, cancer = "all")
}
\arguments{
\item{cdm}{A cdm instance.}

\item{path}{The character name of the folder where the concept
sets are saved within the study package, default \code{"concept_sets"}.}

\item{name}{The character name of the new cohort table to be created.}

\item{cancer}{A character vector of cancer types to include in the cohort, default \code{"all"}. Other possible values are
\code{"bladder_cancer"}, \code{"breast_cancer"}, \code{"colorectal_cancer"}, \code{"lung_cancer"}, \code{"melanoma_of_skin"}, \code{"oesophageal_cancer"}, \code{"prostate_cancer"}.}
}
\value{
A cdm instance including with the newly created cohort table.
}
\description{
The function \code{createCancerCohorts} relies on the \code{CohortConstructor} package to generate cohorts
based on one or more concept sets and additional phenotype specifications.
In this study there are seven cancer types of interest: bladder cancer (1937),
breast cancer (4415), colorectal cancer (1444), esophageal cancer (1414),
lung cancer (1413), prostate cancer (1489) and skin melanoma (1451).
To each cancer type are applied the following phenotype specifications:
\itemize{
\item Study period: from 01-01-2010 to the end of observation period.
\item Inclusion criteria:
\enumerate{
\item The date of the first diagnosis of cancer must fall within the study period.
\item The patient's age must be \eqn{\geq 18} at the time of cancer diagnosis.
}
\item Exclusion criteria:
\enumerate{
\item Exclude patients with death before and at index date.
\item Exclude males from patients with breast cancer.
\item Exclude females from patients with prostate cancer.
}
}
}
