% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.R
\name{createCharacteristicsCohorts}
\alias{createCharacteristicsCohorts}
\title{Creates cancer-related characteristics cohorts}
\usage{
createCharacteristicsCohorts(
  cdm,
  path = "concept_sets",
  name = "cancer_characteristics",
  characteristics = "all"
)
}
\arguments{
\item{cdm}{A cdm instance.}

\item{path}{The character name of the folder where the concept
sets are saved within the study package, default \code{"concept_sets"}.}

\item{name}{The character name of the new cohort table to be created.}

\item{characteristics}{A character vector of cancer-related characteristics to include in the cohort, default \code{"all"}. Other possible values are
"biomarkers", "cancer_cohorts_deck", "cancer_progression", "grade", "radiotherapy", "stage", "surgery" and "treatments_procedures"}
}
\value{
A cdm instance including with the newly created cohort table.
}
\description{
The function \code{createCharacteristicsCohorts} relies on the \code{CohortConstructor} package to generate cohorts
based on one or more concept sets and additional phenotype specifications.
In this study the cancer-related characteristics that will be analysed are: cancer-related,
characteristics (i.e. cancer stage and grade, cancer-specific biomarkers, performance status,
cancer treatments, and laboratory tests and procedures)
}
