% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveStagesRules.R
\name{saveStagesRules}
\alias{saveStagesRules}
\title{Save TNM staging rules to an RDS file}
\usage{
saveStagesRules(
  path = here::here("extras"),
  results_path = system.file("tnm_files", package = "oncomop")
)
}
\arguments{
\item{path}{Character directory where the original .csv files are stored.}

\item{results_path}{Character directory where the RDS files are to be saved.}
}
\value{
\code{NULL}, called for its side effects.
}
\description{
The function reads the \code{.csv} files containing the TNM rules to derive the summary stage.
}
\details{
The files that we refer to are:
\itemize{
\item \code{tnm_concepts}: contains the concept ids of each TNM component differentiated by type
and edition of the UICC classification system;
\item \code{tnm_mapping}: contains the complete rules to map all combinations of TNM components to a
summary stage I-IV, differentiated by edition of staging system and by cancer type.
}
}
