% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRules.R
\name{saveSubtypeRules}
\alias{saveSubtypeRules}
\title{\code{saveSubtypeRules()} to an RDS file}
\usage{
saveSubtypeRules(
  path = here::here("extras"),
  results_path = system.file("data", package = "oncomop")
)
}
\arguments{
\item{path}{Character directory where the original .csv files are stored.}

\item{results_path}{Character directory where the RDS files are to be saved.}
}
\value{
\code{NULL}, called for its side effects.
}
\description{
The function reads the \code{.csv} files containing the subtype rules to derive the summary stage.
}
\details{
The files are:
\itemize{
\item \code{subtype_mapping}: contains the complete rules to map breast cancer subtypes
}
}
