% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagesRules.R
\name{stagesRules}
\alias{stagesRules}
\title{\code{stagesRuleset()} extracts and filters specific set of rules for every cancer}
\usage{
stagesRules(edition, cancer, type)
}
\arguments{
\item{edition}{A choice of "unspecified", "7th" and "8th".}

\item{cancer}{In character, the affected site, a choice of:
"bladder", "breast", "colorectal", "lung", "melanoma", "oesophagus"
and "prostate".}

\item{type}{A choice from "base", "clinical" or "pathological" stage rule.}
}
\value{
\code{NULL}, called for its side effects.
}
\description{
\code{stagesRuleset()} extracts and filters specific set of rules for every cancer
}
