% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDuplicates.R
\name{vizConceptDuplicates}
\alias{vizConceptDuplicates}
\title{Visualize static UpSet plot of staging codelists with duplicates}
\usage{
vizConceptDuplicates(
  cdm,
  concept_folder = "stages",
  input_subcategory = NULL,
  input_edition = NULL,
  input_classification = NULL
)
}
\arguments{
\item{cdm}{A cdm instance, needed to extract concept sets.}

\item{concept_folder}{The character string indicating the folder under \code{inst/concept_sets}
where the concept sets of interest are saved in json files, default \code{"stages"}.}

\item{input_subcategory}{The specific subcategory to filter by (\code{"T0"}, \code{"N0"}, \code{"M0"}, \code{...}),
default \code{NULL} (corresponds to all subcategories).}

\item{input_edition}{The staging system edition to filter by (\code{"7th"}, \code{"8th"}, \verb{"unspecified}),
default \code{NULL} (corresponds to all editions).}

\item{input_classification}{The classification type to filter by (\code{"clinical"}, \code{"pathological"},
\code{"unspecified"}), default \code{NULL} (corresponds to all classifications).}
}
\value{
Prints the plot and returns \code{NULL}.
}
\description{
An UpSet plot is a good alternative to a Venn diagram to visualize intersections
of more than \eqn{3} sets: it shows which sets have common elements with which other set
and the size of the intersection, as well as the original size of every set.

The function \code{vizConceptDuplicates()} allows to filter available staging codelists by subcategory,
edition or classification and plots the intersections of codelists of interest.
}
